Use consolidated event parser and fix test failure
